requirements.txt
pystray
Pillow
pyshortcuts